Must reset pop-up menu value when changing items (for qt)
This was a very, very bad bug. Whenever a plug-in is added or removed
in Hadron, any existing HadronModTargetControls update the pop-up menu
with the new list of plug-ins. In cocoa and swing, this would not
obliterate the pop-up menu's value, but in qt, it does. This corrupted
the state of the target control, and if you saved the Hadron patch,
the resulting file would not load.

It's also an object lesson in NOT using GUI objects to handle state.
Bad, bad, BAD design choice by Batuhan. GUI objects should represent
values that are held elsewhere. They should not serve as data storage.