Remove badvalue synth optimization for main output busses (keep others)
One plugin's outputs may be routed to different other plugins --
so it was not a safe assumption that the mainOutBusses in the
source plugin would be contiguous. Now each item of mainOutBusses
is sent as a separate argument, and separate Out units are created
for each.

inBusses and outBusses are private to the plugin and can be
assumed to be contiguous.