Proper HrRangeMap fix: avoid confusion between input and output mapping
Previous fix was incorrect. The real bug is that I was using isMapped --
which is true if the range map is feeding its output TO another plug-in's
modulation target -- to decide whether the range map's modValue modulation
input is receiving data FROM another plug-in. The right condition is
mappedMods[\modValue].notNil.

Now doWakeFromLoad can be done in the normal place and it all works.